# api.yaml
openapi: 3.0.0
info:
  title: Vehicle Parking App API
  version: 1.0.0
  description: API documentation for authentication, admin, and user operations in the Vehicle Parking App
...
paths:
...
  /admin/addlot:
    get:
      tags: [Admin]
      summary: Render add parking lot page
      responses:
        200:
          description: Add lot HTML page
    post:
      tags: [Admin]
      summary: Add a new parking lot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                locationName: { type: string }
                address: { type: string }
                pincode: { type: string }
                pricePerHour: { type: string }
                maxSpots: { type: integer }
              required: [locationName, address, pincode, pricePerHour, maxSpots]
      responses:
        200:
          description: Parking Lot added successfully (JSON)
        302:
          description: Redirect on error
  /admin/edit:
    get:
      tags: [Admin]
      summary: Render edit parking lot page
      parameters:
        - in: query
          name: lot_id
          schema: { type: integer }
          required: true
      responses:
        200:
          description: Edit parking lot HTML page
    post:
      tags: [Admin]
      summary: Edit parking lot details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                locationName: { type: string }
                address: { type: string }
                pincode: { type: string }
                pricePerHour: { type: string }
                maxSpots: { type: integer }
                lot_id: { type: integer }
              required: [locationName, address, pincode, pricePerHour, maxSpots, lot_id]
      responses:
        200:
          description: Parking lot updated successfully (JSON)
        400:
          description: Missing required fields
  /admin/delete:
    get:
      tags: [Admin]
      summary: Render confirm delete parking lot page
      parameters:
        - in: query
          name: lot_id
          schema: { type: integer }
          required: true
      responses:
        200:
          description: Confirm delete HTML page
    post:
      tags: [Admin]
      summary: Delete a parking lot and its spots
      parameters:
        - in: query
          name: lot_id
          schema: { type: integer }
          required: true
      responses:
        302:
          description: Redirect to dashboard after deletion
  /admin/viewSpot:
    get:
      tags: [Admin]
      summary: Render view spot page
      parameters:
        - in: query
          name: spot_id
          schema: { type: integer }
          required: true
        - in: query
          name: lot_id
          schema: { type: integer }
          required: true
        - in: query
          name: status
          schema: { type: string }
          required: true
      responses:
        200:
          description: View spot HTML page
    post:
      tags: [Admin]
      summary: Delete a particular parking spot
      parameters:
        - in: query
          name: spot_id
          schema: { type: integer }
          required: true
        - in: query
          name: lot_id
          schema: { type: integer }
          required: true
        - in: query
          name: status
          schema: { type: string }
          required: true
      responses:
        302:
          description: Redirect to dashboard after deletion
...